refactor: moved checksumprovider from zeebe-db to snapshots
(cherry picked from commit babd4a51aa53673765289bdeefd8a3a9cc409e3d)
