+++ restore/pom.xml
@@ -103,10 +103,6 @@
             <!-- javac and maven-dependency-plugin disagree whether this is needed or not. -->
             <ignoredNonTestScopedDependency>io.camunda:zeebe-scheduler</ignoredNonTestScopedDependency>
           </ignoredNonTestScopedDependencies>
-          <!-- This is transitive dependency-->
-          <ignoredUsedUndeclaredDependencies>
-            <ignoredUsedUndeclaredDependency>io.camunda:zeebe-db</ignoredUsedUndeclaredDependency>
-          </ignoredUsedUndeclaredDependencies>
         </configuration>
       </plugin>
     </plugins>
